requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
brython==3.11.3
certifi==2024.2.2
click==8.1.7
elastic-transport==8.12.0
elasticsearch==8.12.0
fastapi==0.109.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
loguru==0.7.2
magic-filter==1.0.12
multidict==6.0.5
pyaes==1.6.1
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
PySocks==1.7.1
python-multipart==0.0.9
rsa==4.9
sniffio==1.3.0
socks==0
SQLAlchemy==2.0.25
starlette==0.35.1
Telethon==1.33.1
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.0
websockets==12.0
yarl==1.9.4
jinja2==3.1.4
email_validator==2.2.0
python-dotenv~=1.0.1
pytz~=2024.1
phonenumbers~=8.13.40
python-jose==3.3.0
bcrypt